Update so pixel watch bedtime sync works
I *think* all that's needed to make the pixel watch bedtime sync work is to edit the midx value (which is currently the "middle" of the row for samsung watches) to instead be divided by 3. That way we get the first (left) 3rd of the screen as the click target for the midx value. And that's where the pixel watch has the bedtime mode button. I was too lazy to install the android sdk and build/debug this myself - hoping the dev is willing to give it a shot for us :).